terra_mars/requirements.txt
celery==4.3.0
django==2.2.6
djangorestframework==3.10.3
psycopg2>=2.7,<3.0
redis==3.3.11
